Generate object variants for oneofs
This fixes a segfault problem that I had, when running the conformance test
suite. The following piece of code segfaulted on the last assignment:

	msg.oneofUint32 = 1
	msg.oneofString = "test"

If the first assignment was `0`, then there was no problem.

Now we use the object variant, which gives nim runtime the ability to track
what data the union holds.