Fix Duration JSON parsing
I use `include duration_pb` in the `duration.nim` file. This has the effect of
not using the property setter procs when doing assignments! Instead, we directly
use the member variable bypassing the setter. This means that `hasField` is not
updated. Which means that when serializing a message that was parsed from JSON,
we don't serialize the `Duration.seconds` and `Duration.nanos` fields because
the library thinks they are not set.

The fix for now is to use the dedicated `setFoo` procs for setting fields in
these WKT extra functionality files which include the corresponding `*_pb.nim`
file.

This is not a problem when using `import`, as the fields are private to the
`*_pb.nim` module, so outside code uses the property setters when doing
assignments.